requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==2.0.12
Django==4.2.14
django-filter==24.2
django-jazzmin==3.0.0
idna==3.7
mysql-cli
pillow==10.4.0
python-dotenv==1.0.1
requests==2.26.0
sqlparse==0.5.1
sslcommerz-lib==1.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.26.19
whitenoise==6.7.0
wrapt==1.12.1
